Fixed some commit-related issues. Implemented Undo stack handling.
Detailed changes:
- modify.push() no longer takes the level as argument; instead it is
extracted from the UndoCommand's level attribute.
- Since the redo of a REAL command while EDITOR is the active stack
clears the editor stack, undoing a commit now clears all editors.
Ideally, they should be restored on "redo", which doesn't work yet.
- the fileTags attribute is now set on import by the editors and no
longer by the NewElementsCommand
- the undo stack now has a method setState(level). If the current level
is EDITOR and setState(REAL) is called, a question dialog is shown and
reminds the user that the editor stack will be lost. If this is
rejected, an exception is thrown. Thus methods calling push() and
beginMacro() with REAL commands should handle this exception.