Started to implement TreeView's new context menu framework.
Subclasses of TreeView have a classmethod contextMenu() returning
a list that defines the layout of the part of the context menu
coming from that class. The list contains either TreeActions or
named lists (i.e. lists with a 'name' attribute) representing
submenus.
If a subclass wants to implement an own context menu part, it should
override the classmethod initContextMenu() that returns a list of
the above type.
Plugins may use the contextMenu() method to extend the menu.