Preliminary support for different "file" backends.
A file is now the general representation of something that can be
played, and might also be a stream, a track on a disc, etc. It is
identified by an URL which supersedes the old *path* attribute. Plugins
can provide additional protocols; normal files are now identified by the
URL file:///relative/path/to/music/directory.