Added mpd:// URL type for files MPD knows but we don't.
If the connected MPD's playlist contains a file that is not contained in
the local music directory, an MPDFile with an MPDURL of the form
mpd//profilename/path is created. Its tags are read using the MPD
protocol.

Alongside I have adapted the playlist to use URLs instead of paths.