Fix Bug #130.
The QSlider class does not support jumping to a slider position directly. For the moment I worked around this by using QStyle.sliderValueFromPosition. For OMG 0.4 the playback widget should be rewritten anyway.

In Phonon, the slider was disabled because setCurrent() in PhononPlayerBackend called play() instead of setState(PLAY), which circumveited the stateChanged signal.