Fixes #10 and changes the handling of "special" tags.
BackendFile now creates an OrderedDict "specialTags" for tags like
tracknumber, discnumber and compilation (perhaps more in the future).

The album guesser now can be configured to give albums which have the
compilation tag set (to something different than "0") a specific flag.

Also a lot of bugs in the album guesser were fixed; most importantly
changes to the profile were never saved in the current version. :)