Experimental support for web radio.
There's a new backend file type, HTTPStream, with an associated URL type
HTTPStreamURL, which can be used for Web radios or other HTTP streams.
Currently the only method to add such a stream is a button in the editor
gui which is not meant to stay there forever. :-) Seems to work with
both phonon and MPD backend.

In that context I have fixed several bugs with non-realfile URLs and
backend files.