Implement a UnknownURL scheme. See issue #163.
This at least fixes the browser crashing. However there's a small issue
left: when the according plugin, e.g. audiocd, is loaded after an
UnknownURL has been created, it should be converted to the real thing.

Perhaps we could guard filebackends.urlTypes by add and remove functions
taking care of this. Does it suffice to go through all levels and simply
change the URLs of all files with an UnknownURL of the correct type?