Warning: This commit requires a database change.
Diacritics-insensitive search (the default) is now performed using a special column 'search_value' in values_varchar. This works the same with MySQL and SQLite and is much faster than the previous SQLite-implementation. A check has been added to the DB-Analyzer find and fix incorrect search values.

To change the database simply add the column search_value to values_varchar (see the definition in database.tables) and run the check and fix in the DB-Analyzer.